[DF-10551] Compilation fixes
Former-commit-id: 0d61d18140a90be02670961f061f770968bc90d5

+++ Sources/Internal/Reflection/Private/Reflection.cpp
@@ -119,7 +119,7 @@ struct Dumper
 
             line << "(";
 
-            if (::strlen(typeName) > w)
+            if (::strlen(typeName) > static_cast<size_t>(w))
             {
                 line.write(typeName, w);
             }
